# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "You can visit out RiderTrack page at [https://rider-track-dev.herokuapp.com]" 
  version: "1.0.0"
  title: "Ridertrack Swagger"
host: "rider-track-dev.herokuapp.com"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /events:
    get:
      tags:
      - "event"
      summary: "Finds all events"
      description: "It returns the list of all the events. It accepts query params for filtering events: country, type"
      produces:
        - application/json
      parameters:
      - name: "country"
        in: "query"
        description: "Filter by country"
        type: "string"
      - name: "type"
        in: "query"
        description: "Filter by type"
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/createEvent'
        400:
          description: "Error message"
    post:
      tags:
      - "event"
      summary: "Creates a new event"
      description: "creates a new event with given parameters"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "Event object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/createEvent"
      responses:
        400:
          description: "errors"
        200:
          description: "Event successfully created"
  /events/{eventId}:
    get:
      tags:
      - "event"
      summary: "Find a event by given Id"
      description: "Find a event by given Id"
      produces:
        - application/json
      parameters:
      - in: "path"
        name: "eventId"
        type: "string"
        description: "The id of the event you want to get"
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/Event'
        400:
          description: "Error message"
    put:
      tags:
      - "event"
      summary: "Update a event by given Id"
      description: "Update the data of a event by given Id"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        description: "The id of the event you want to update"
        required: true
      - in: body
        name: body
        description: "Event object with parameters that you want to update"
        required: true
        schema:
          $ref: "#/definitions/Event"
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/Event'
        400:
          description: "Error message"
        401:
          description: "You are not allowed to change event"
    delete:
      tags:
      - "event"
      summary: "Delete a event"
      description: "Delete a event by given Id"
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        description: "The id of the event you want to delete."
        required: true
      responses:
        200:
          description: "Event successfully deleted"
        400:
          description: "Error message"
        401:
          description: "You are not allowed to delete this event"
...
  /events/{eventId}/participantsList:
    get:
      tags:
      - "event"
      summary: "Find all participant of an event "
      description: "Find all participants of an event by given Id"
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        description: "The id of the event of which you want to see participants list."
        required: true
      responses:
        200:
          description: "{participants: id1, id2, id3}"
        400:
          description: "Error message"
  /events/{eventId}/logo:
    get:
      tags:
      - "event"
      summary: "Find the logo of an event "
      description: "Find the logo of an event by given Id"
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        description: "The id of the event of which you want to see the logo."
        required: true
      responses:
        200:
          description: "*showing the logo of event"
        400:
          description: "Error message"
  /events/{eventId}/organizer:
    get:
      tags:
      - "event"
      summary: "Find an organizer of an event"
      description: "Finds an organizer of an event"
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        description: "The id of the event of which you want to see who organized it."
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/User'
        400:
          description: "Error message"
  /events/{eventId}/participants/positions:
    get:
      tags:
      - "event"
      summary: "Finds last positions of all participants"
      description: "It returns a list with the most recent positions of all the users in the event."
      produces:
        - application/json
      parameters:
      - name: eventId
        in: path
        type: "string"
        required: true
      responses:
        200:
          description: "list of positions"
          schema:
            $ref: "#/definitions/arrayOfPositions"
        400:
          description: "Error message"
    post:
      tags:
      - "event"
      summary: "Adding a position"
      description: "Gets the location from a mobile app in form of an object[lat,lon]. Adds the coordinates(String) of last location to the end of location list. If a location for an user in an event isn't created it creates it, if it is created than it updates it with adding the coordinates(String)."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: eventId
        in: path
        type: string
        required: true
      - in: body
        name: body
        description: "Position object that is going to be added"
        required: true
        schema:
          description: ""
      responses:
        400:
          description: "errors"
        200:
          description: "Positions updated successfully"
          schema:
            $ref: "#/definitions/Position"
  /events/{eventId}/{userId}/location:
    get:
      tags:
      - "event"
      summary: "Last location of an user"
      description: "Gets the last location of an user."
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        required: true
      - in: path
        name: "userId"
        type: "string"
        description: "The id of the user of which you want to know location."
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/lastPosition'
        400:
          description: "Error message"
  /events/{eventId}/ranking:
    get:
      tags:
      - "event"
      summary: "Latest ranking of an event"
      description: "This method gets the latest ranking of the event"
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        required: true
      responses:
        200:
          description: "Ranking here : ranking was sent on: "
        400:
          description: "Error message"
  /events/{eventId}/route:
    get:
      tags:
      - "route"
      summary: "It returns the route of the requested event."
      description: "It returns the route of the requested event."
      produces:
      - "application/json"
      parameters:
      - name: eventId
        in: path
        type: "string"
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/Route"
        400:
          description: "Error message"
    post:
      tags:
      - "route"
      summary: "It creates the route for the given event."
      description: "It creates the route for the given event."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: eventId
        type: "string"
        required: true
      - in: body
        name: body
        description: "Object of coordinates that make a route"
        required: true
        schema:
          description: ""
      responses:
        400:
          description: "errors"
        200:
          description: "Route successfully created"
          schema:
            $ref: "#/definitions/Coordinates"
    put:
      tags:
      - "route"
      summary: "Update route of an event"
      description: "It updates the route for the given event."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: eventId
        in: path
        type: string
        required: true
      - in: body
        name: body
        description: "Object of coordinates that make a route"
        required: true
        schema:
          description: ""
      responses:
        200:
          description: "Route successfully updated"
          schema:
            $ref: "#/definitions/Coordinates"
        400:
          description: "Error message"
    delete:
      tags:
      - "route"
      summary: "Delete a route"
      description: "Delete route of an event by given Id"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        required: true
      responses:
        200:
          description: "Route successfully deleted"
        400:
          description: "Error message"
  /events/{eventId}/tracking/start:
    post:
      tags:
      - "tracking"
      summary: "It starts the tracking of an event."
      description: "It starts the tracking of an event. It uses the method on the event model."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        type: "string"
        required: true
      responses:
        400:
          description: "errors"
        200:
          description: "Tracking started successfully."
  /events/{eventId}/tracking/stop:
    post:
      tags:
      - "tracking"
      summary: "It stops the tracking of an event."
      description: "It stops the tracking of an event. It uses the method on the event model."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        type: "string"
        required: true
      responses:
        400:
          description: "errors"
        200:
          description: "Tracking stopped successfully."
...
  /admin/events:
    get:
      tags:
      - "admin"
      summary: "Finds all events"
      description: "It returns the list of all the events. It accepts query params for filtering events: country, type"
      produces:
        - application/json
      parameters:
      - name: "country"
        in: "query"
        description: "Filter by country"
        type: "string"
      - name: "type"
        in: "query"
        description: "Filter by type"
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/createEvent'
        400:
          description: "Error message"
    post:
      tags:
      - "event"
      summary: "Creates a new event"
      description: "creates a new event with given parameters"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: body
        name: body
        description: "Event object that needs to be added"
        required: true
        schema:
          $ref: "#/definitions/createEvent"
      responses:
        400:
          description: "errors"
        200:
          description: "Event successfully created"
...
  /admin/events/{eventId}/participantsList:
    get:
      tags:
      - admin
      summary: "Find all participant of an event "
      description: "Find all participants of an event by given Id"
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        description: "The id of the event of which you want to see participants list."
        required: true
      responses:
        200:
          description: "{participants: id1, id2, id3}"
        400:
          description: "Error message"
  /admin/events/{eventId}/logo:
    get:
      tags:
      - admin
      summary: "Find the logo of an event "
      description: "Find the logo of an event by given Id"
      produces:
        - application/json
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        description: "The id of the event of which you want to see the logo."
        required: true
      responses:
        200:
          description: "*showing the logo of event"
        400:
          description: "Error message"
...
definitions:
  User:
    type: "object"
    properties:
      _id:
        type: "string"
      email:
        type: "string"
      name:
        type: "string"
      surname:
        type: "string"
      role:
        type: "string"
        enum:
        - "user"
        - "administrator"
...
  Event:
    type: "object"
    properties:
      _id:
        type: "string"
      organizerId:
        type: "string"
      name:
        type: "string"
      country:
        type: "string"
      city:
        type: "string"
      startingDate:
        type: "string"
      type:
        type: "string"
        enum:
        - "running"
        - "cycling"
        - "hiking"
        - "triathlon"
        - "other"
  createEvent:
     type: "object"
     required:
     - "organizerId"
     - "name"
     - "country"
     - "city"
     - "startingDate"
     - "type"
     properties:
       organizerId:
         type: "string"
       name:
         type: "string"
       country:
         type: "string"
       city:
         type: "string"
       startingDate:
         type: "string"
       type:
         type: "string"
         enum:
         - "running"
         - "cycling"
         - "hiking"
         - "triathlon"
         - "other"
  Position:
    type: "object"
    properties:
      eventId:
        type: "string"
      userId:
        type: "string"
      lastPosition:
        $ref: "#/definitions/lastPosition"
...
  arrayOfPositions:
    type: "object"
    properties:
      positions:
        type: "array"
        items:
          $ref: "#/definitions/Position"
...